requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
aiomysql
python-jose
passlib[bcrypt]
python-multipart
pytest
httpx
pytest-cov
factory-boy

